Docker & Docker Compose configurations fixed & upgraded.
