+++ .dockerignore
@@ -1,0 +1,12 @@
+start_endpoint.bat
+start.bat
+.gitignore
+.github
+.git
+.env
+LICENCE
+install_ubuntu.sh
+install_on_huggingface.md
+docker-compose.yml
+config.json
+docker-compose/
