--- docker-compose.yml
@@ -1,10 +1,0 @@
-version: "3.10"
-
-services:
-  neurogengpt:
-    build: .
-    ports:
-      - "7860:7860"
-      - "1337:1337"
-    env_file:
-      - .env
