+++ docker-compose/.env
@@ -1,0 +1,9 @@
+LANG=ru
+# Paste ur email below
+EMAIL=acme@example.com
+TRAEFIK_USER=admin
+CERT_RESOLVER=letsencrypt
+# To change password use " htpasswd -nBC 10 admin " (or whatever username u want) and paste below the part after :
+TRAEFIK_PASSWORD_HASH=
+OPENAI_API_KEY=
+DOMAIN=
