+++ docker-compose/docker-compose.yml
@@ -1,0 +1,14 @@
+version: "3.10"
+
+services:
+  neurogengpt:
+    build:
+      context: ../
+      dockerfile: Dockerfile
+    ports:
+      - "7860:7860"
+      - "1337:1337"
+    volumes:
+      - ../config.json:/app/config.json
+    env_file:
+      - .env
