+++ dockerfile
@@ -1,18 +1,28 @@
-FROM python:3.10
+# Stage 1: Build
+
+FROM python:3.10-slim-buster as builder
 
 WORKDIR /app
 
-COPY . .
+RUN apt-get update && apt-get install -y --no-install-recommends build-essential git
+COPY requirements.txt .
+RUN pip install --no-cache-dir --user -r requirements.txt
 
-RUN pip install --upgrade pip 
-RUN pip install -r requirements.txt
-RUN pip install --upgrade pip && \
-    pip install -r requirements.txt && \
-    python -m spacy download ru_core_news_sm  && \
+FROM python:3.10-slim-buster
+
+WORKDIR /app
+
+COPY --from=builder /root/.local /root/.local
+ENV PATH=/root/.local/bin:$PATH
+COPY . .
+RUN python -m spacy download ru_core_news_sm && \
     python -m spacy download en_core_web_sm && \
     python -m spacy download zh_core_web_sm
+RUN chmod +x entrypoint.sh
 
 EXPOSE 7860
 EXPOSE 1337
 
-CMD ["python", "webui_ru.py"]
+ENV LANG=ru
+
+CMD ["./entrypoint.sh"]
