+++ testing/test_providers.py
@@ -1,76 +1,69 @@
+import random, string
 import sys
 from pathlib import Path
-from colorama import Fore
 
 sys.path.append(str(Path(__file__).parent.parent))
 
 from g4f import BaseProvider, models, Provider
 
-logging = False
 
 def main():
     providers = get_providers()
-    failed_providers = []
+    results: list[list[str | bool]] = []
 
     for _provider in providers:
-        if _provider.needs_auth:
-            continue
-        print("Provider:", _provider.__name__)
-        result = test(_provider)
-        print("Result:", result)
-        if _provider.working and not result:
-            failed_providers.append(_provider)
+        print("start", _provider.__name__)
+        actual_working = judge(_provider)
+        expected_working = _provider.working
+        match = actual_working == expected_working
 
-    print()
+        results.append([_provider.__name__, expected_working, actual_working, match])
 
-    if failed_providers:
-        print(f"{Fore.RED}Failed providers:\n")
-        for _provider in failed_providers:
-            print(f"{Fore.RED}{_provider.__name__}")
-    else:
-        print(f"{Fore.GREEN}All providers are working")
+    print("failed provider list")
+    for result in results:
+        if not result[3]:
+            print(result)
 
 
 def get_providers() -> list[type[BaseProvider]]:
     provider_names = dir(Provider)
     ignore_names = [
         "base_provider",
-        "BaseProvider"
+        "BaseProvider",
     ]
     provider_names = [
         provider_name
         for provider_name in provider_names
         if not provider_name.startswith("__") and provider_name not in ignore_names
     ]
-    return [getattr(Provider, provider_name) for provider_name in sorted(provider_names)]
+    return [getattr(Provider, provider_name) for provider_name in provider_names]
 
 
-def create_response(_provider: type[BaseProvider]) -> str:
-    if _provider.supports_gpt_35_turbo:
-        model = models.gpt_35_turbo.name    
-    elif _provider.supports_gpt_4:
-        model = models.gpt_4
-    elif hasattr(_provider, "model"):
-        model = _provider.model
-    else:
-        model = None
+def create_response(_provider: type[BaseProvider], _str: str) -> str:
+    model = (
+        models.gpt_35_turbo.name
+        if _provider is not Provider.H2o
+        else models.falcon_7b.name
+    )
     response = _provider.create_completion(
         model=model,
-        messages=[{"role": "user", "content": "Hello"}],
+        messages=[{"role": "user", "content": f"just output \"{_str}\""}],
         stream=False,
     )
     return "".join(response)
 
-    
-def test(_provider: type[BaseProvider]) -> bool:
+
+def judge(_provider: type[BaseProvider]) -> bool:
+    if _provider.needs_auth:
+        return _provider.working
+
     try:
-        response = create_response(_provider)
+        _str = "".join(random.choices(string.ascii_letters + string.digits, k=4))
+        response = create_response(_provider, _str)
         assert type(response) is str
-        assert len(response) > 0
-        return response
+        return len(response) > 1 and _str in response
     except Exception as e:
-        if logging:
-            print(e)
+        print(f"{_provider.__name__}: {str(e)}")
         return False
